Split workers out of supervisor module
