+++ relayer/src/lib.rs
@@ -24,9 +24,11 @@ pub mod keyring;
 pub mod light_client;
 pub mod link;
 pub mod macros;
+pub mod object;
 pub mod registry;
 pub mod relay;
 pub mod supervisor;
 pub mod transfer;
 pub mod upgrade_chain;
 pub mod util;
+pub mod worker;
