+++ relayer/src/object.rs
@@ -1,0 +1,243 @@
+use anomaly::BoxError;
+
+use ibc::{
+    ics02_client::{client_state::ClientState, events::UpdateClient},
+    ics04_channel::events::{
+        Attributes, CloseInit, SendPacket, TimeoutPacket, WriteAcknowledgement,
+    },
+    ics24_host::identifier::{ChainId, ChannelId, ClientId, PortId},
+    Height,
+};
+
+use crate::chain::{
+    counterparty::{channel_connection_client, get_counterparty_chain},
+    handle::ChainHandle,
+};
+
+/// Client
+#[derive(Clone, Debug, PartialEq, Eq, Hash)]
+pub struct Client {
+    /// Destination chain identifier.
+    pub dst_chain_id: ChainId,
+
+    /// Source channel identifier.
+    pub dst_client_id: ClientId,
+
+    /// Source chain identifier.
+    pub src_chain_id: ChainId,
+}
+
+impl Client {
+    pub fn short_name(&self) -> String {
+        format!(
+            "{}->{}:{}",
+            self.src_chain_id, self.dst_chain_id, self.dst_client_id
+        )
+    }
+}
+
+/// A unidirectional path from a source chain, channel and port.
+#[derive(Clone, Debug, PartialEq, Eq, Hash)]
+pub struct UnidirectionalChannelPath {
+    /// Destination chain identifier.
+    pub dst_chain_id: ChainId,
+
+    /// Source chain identifier.
+    pub src_chain_id: ChainId,
+
+    /// Source channel identifier.
+    pub src_channel_id: ChannelId,
+
+    /// Source port identifier.
+    pub src_port_id: PortId,
+}
+
+impl UnidirectionalChannelPath {
+    pub fn short_name(&self) -> String {
+        format!(
+            "{}/{}:{}->{}",
+            self.src_channel_id, self.src_port_id, self.src_chain_id, self.dst_chain_id,
+        )
+    }
+}
+
+/// An object determines the amount of parallelism that can
+/// be exercised when processing [`IbcEvent`] between
+/// two chains. For each [`Object`], a corresponding
+/// [`Worker`] is spawned and all [`IbcEvent`]s mapped
+/// to an [`Object`] are sent to the associated [`Worker`]
+/// for processing.
+#[derive(Clone, Debug, PartialEq, Eq, Hash)]
+pub enum Object {
+    /// See [`Client`].
+    Client(Client),
+    /// See [`UnidirectionalChannelPath`].
+    UnidirectionalChannelPath(UnidirectionalChannelPath),
+}
+
+impl Object {
+    /// Returns `true` if this [`Object`] is for a [`Worker`] which is interested
+    /// in new block events originating from the chain with the given [`ChainId`].
+    /// Returns `false` otherwise.
+    pub fn notify_new_block(&self, src_chain_id: &ChainId) -> bool {
+        match self {
+            Object::UnidirectionalChannelPath(p) => p.src_chain_id == *src_chain_id,
+            Object::Client(_) => false,
+        }
+    }
+}
+
+impl From<Client> for Object {
+    fn from(c: Client) -> Self {
+        Self::Client(c)
+    }
+}
+
+impl From<UnidirectionalChannelPath> for Object {
+    fn from(p: UnidirectionalChannelPath) -> Self {
+        Self::UnidirectionalChannelPath(p)
+    }
+}
+
+impl Object {
+    pub fn src_chain_id(&self) -> &ChainId {
+        match self {
+            Self::Client(ref client) => &client.src_chain_id,
+            Self::UnidirectionalChannelPath(ref path) => &path.src_chain_id,
+        }
+    }
+
+    pub fn dst_chain_id(&self) -> &ChainId {
+        match self {
+            Self::Client(ref client) => &client.dst_chain_id,
+            Self::UnidirectionalChannelPath(ref path) => &path.dst_chain_id,
+        }
+    }
+
+    pub fn short_name(&self) -> String {
+        match self {
+            Self::Client(ref client) => client.short_name(),
+            Self::UnidirectionalChannelPath(ref path) => path.short_name(),
+        }
+    }
+
+    /// Build the object associated with the given [`UpdateClient`] event.
+    pub fn for_update_client(
+        e: &UpdateClient,
+        dst_chain: &dyn ChainHandle,
+    ) -> Result<Self, BoxError> {
+        let client_state = dst_chain.query_client_state(e.client_id(), Height::zero())?;
+        if client_state.refresh_period().is_none() {
+            return Err(format!(
+                "client '{}' on chain {} does not require refresh",
+                e.client_id(),
+                dst_chain.id()
+            )
+            .into());
+        }
+
+        let src_chain_id = client_state.chain_id();
+
+        Ok(Client {
+            dst_client_id: e.client_id().clone(),
+            dst_chain_id: dst_chain.id(),
+            src_chain_id,
+        }
+        .into())
+    }
+
+    /// Build the client object associated with the given channel event attributes.
+    pub fn for_chan_open_events(
+        e: &Attributes,
+        dst_chain: &dyn ChainHandle,
+    ) -> Result<Self, BoxError> {
+        let channel_id = e
+            .channel_id()
+            .as_ref()
+            .ok_or_else(|| format!("channel_id missing in OpenAck event '{:?}'", e))?;
+
+        let client = channel_connection_client(dst_chain, e.port_id(), channel_id)?.client;
+        if client.client_state.refresh_period().is_none() {
+            return Err(format!(
+                "client '{}' on chain {} does not require refresh",
+                client.client_id,
+                dst_chain.id()
+            )
+            .into());
+        }
+
+        Ok(Client {
+            dst_client_id: client.client_id.clone(),
+            dst_chain_id: dst_chain.id(),
+            src_chain_id: client.client_state.chain_id(),
+        }
+        .into())
+    }
+
+    /// Build the object associated with the given [`SendPacket`] event.
+    pub fn for_send_packet(e: &SendPacket, src_chain: &dyn ChainHandle) -> Result<Self, BoxError> {
+        let dst_chain_id =
+            get_counterparty_chain(src_chain, &e.packet.source_channel, &e.packet.source_port)?;
+
+        Ok(UnidirectionalChannelPath {
+            dst_chain_id,
+            src_chain_id: src_chain.id(),
+            src_channel_id: e.packet.source_channel.clone(),
+            src_port_id: e.packet.source_port.clone(),
+        }
+        .into())
+    }
+
+    /// Build the object associated with the given [`WriteAcknowledgement`] event.
+    pub fn for_write_ack(
+        e: &WriteAcknowledgement,
+        src_chain: &dyn ChainHandle,
+    ) -> Result<Self, BoxError> {
+        let dst_chain_id = get_counterparty_chain(
+            src_chain,
+            &e.packet.destination_channel,
+            &e.packet.destination_port,
+        )?;
+
+        Ok(UnidirectionalChannelPath {
+            dst_chain_id,
+            src_chain_id: src_chain.id(),
+            src_channel_id: e.packet.destination_channel.clone(),
+            src_port_id: e.packet.destination_port.clone(),
+        }
+        .into())
+    }
+
+    /// Build the object associated with the given [`TimeoutPacket`] event.
+    pub fn for_timeout_packet(
+        e: &TimeoutPacket,
+        src_chain: &dyn ChainHandle,
+    ) -> Result<Self, BoxError> {
+        let dst_chain_id =
+            get_counterparty_chain(src_chain, &e.packet.source_channel, &e.packet.source_port)?;
+
+        Ok(UnidirectionalChannelPath {
+            dst_chain_id,
+            src_chain_id: src_chain.id(),
+            src_channel_id: e.src_channel_id().clone(),
+            src_port_id: e.src_port_id().clone(),
+        }
+        .into())
+    }
+
+    /// Build the object associated with the given [`CloseInit`] event.
+    pub fn for_close_init_channel(
+        e: &CloseInit,
+        src_chain: &dyn ChainHandle,
+    ) -> Result<Self, BoxError> {
+        let dst_chain_id = get_counterparty_chain(src_chain, e.channel_id(), &e.port_id())?;
+
+        Ok(UnidirectionalChannelPath {
+            dst_chain_id,
+            src_chain_id: src_chain.id(),
+            src_channel_id: e.channel_id().clone(),
+            src_port_id: e.port_id().clone(),
+        }
+        .into())
+    }
+}
