+++ relayer/src/util.rs
@@ -1,6 +1,9 @@
 mod block_on;
 pub use block_on::block_on;
 
+mod recv_multiple;
+pub use recv_multiple::recv_multiple;
+
 pub mod iter;
 pub mod retry;
 pub mod sled;
