+++ relayer/src/worker.rs
@@ -1,0 +1,192 @@
+use std::{fmt, thread, time::Duration};
+
+use anomaly::BoxError;
+use crossbeam_channel::Receiver;
+
+use tracing::{debug, error, error_span, info, trace, warn};
+
+use ibc::{events::IbcEvent, ics02_client::events::UpdateClient};
+
+use crate::{
+    chain::handle::ChainHandlePair,
+    foreign_client::{ForeignClient, ForeignClientError, MisbehaviourResults},
+    link::{Link, LinkParameters},
+    object::{Client, Object, UnidirectionalChannelPath},
+};
+
+mod handle;
+pub use handle::WorkerHandle;
+
+mod cmd;
+pub use cmd::WorkerCmd;
+
+/// A worker processes batches of events associated with a given [`Object`].
+pub struct Worker {
+    chains: ChainHandlePair,
+    rx: Receiver<WorkerCmd>,
+}
+
+impl fmt::Display for Worker {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "[{} <-> {}]", self.chains.a.id(), self.chains.b.id(),)
+    }
+}
+
+impl Worker {
+    /// Spawn a worker which relay events pertaining to an [`Object`] between two `chains`.
+    pub fn spawn(chains: ChainHandlePair, object: Object) -> WorkerHandle {
+        let (tx, rx) = crossbeam_channel::unbounded();
+
+        debug!(
+            "[{}] spawned worker with chains a:{} and b:{} for object {:#?} ",
+            object.short_name(),
+            chains.a.id(),
+            chains.b.id(),
+            object,
+        );
+
+        let worker = Self { chains, rx };
+        let thread_handle = std::thread::spawn(move || worker.run(object));
+
+        WorkerHandle::new(tx, thread_handle)
+    }
+
+    /// Run the worker event loop.
+    fn run(self, object: Object) {
+        let span = error_span!("worker loop", worker = %object.short_name());
+        let _guard = span.enter();
+
+        let result = match object {
+            Object::UnidirectionalChannelPath(path) => self.run_uni_chan_path(path),
+            Object::Client(client) => self.run_client(client),
+        };
+
+        if let Err(e) = result {
+            error!("worker error: {}", e);
+        }
+
+        info!("worker exits");
+    }
+
+    fn run_client_misbehaviour(
+        &self,
+        client: &ForeignClient,
+        update: Option<UpdateClient>,
+    ) -> bool {
+        match client.detect_misbehaviour_and_submit_evidence(update) {
+            MisbehaviourResults::ValidClient => false,
+            MisbehaviourResults::VerificationError => {
+                // can retry in next call
+                false
+            }
+            MisbehaviourResults::EvidenceSubmitted(_) => {
+                // if evidence was submitted successfully then exit
+                true
+            }
+            MisbehaviourResults::CannotExecute => {
+                // skip misbehaviour checking if chain does not have support for it (i.e. client
+                // update event does not include the header)
+                true
+            }
+        }
+    }
+
+    /// Run the event loop for events associated with a [`Client`].
+    fn run_client(self, client: Client) -> Result<(), BoxError> {
+        let mut client = ForeignClient::restore(
+            &client.dst_client_id,
+            self.chains.a.clone(),
+            self.chains.b.clone(),
+        );
+
+        info!(
+            "running client worker & initial misbehaviour detection for {}",
+            client
+        );
+
+        // initial check for evidence of misbehaviour for all updates
+        let skip_misbehaviour = self.run_client_misbehaviour(&client, None);
+
+        info!(
+            "running client worker (misbehaviour and refresh) for {}",
+            client
+        );
+
+        loop {
+            thread::sleep(Duration::from_millis(600));
+
+            // Run client refresh, exit only if expired or frozen
+            if let Err(e @ ForeignClientError::ExpiredOrFrozen(..)) = client.refresh() {
+                error!("failed to refresh client '{}': {}", client, e);
+                continue;
+            }
+
+            if skip_misbehaviour {
+                continue;
+            }
+
+            if let Ok(WorkerCmd::IbcEvents { batch }) = self.rx.try_recv() {
+                trace!("client '{}' worker receives batch {:?}", client, batch);
+
+                for event in batch.events {
+                    if let IbcEvent::UpdateClient(update) = event {
+                        debug!("client '{}' updated", client);
+
+                        // Run misbehaviour. If evidence submitted the loop will exit in next
+                        // iteration with frozen client
+                        self.run_client_misbehaviour(&client, Some(update));
+                    }
+                }
+            }
+        }
+    }
+
+    /// Run the event loop for events associated with a [`UnidirectionalChannelPath`].
+    fn run_uni_chan_path(self, path: UnidirectionalChannelPath) -> Result<(), BoxError> {
+        let mut link = Link::new_from_opts(
+            self.chains.a.clone(),
+            self.chains.b.clone(),
+            LinkParameters {
+                src_port_id: path.src_port_id,
+                src_channel_id: path.src_channel_id,
+            },
+        )?;
+
+        if link.is_closed()? {
+            warn!("channel is closed, exiting");
+            return Ok(());
+        }
+
+        loop {
+            thread::sleep(Duration::from_millis(200));
+
+            if let Ok(cmd) = self.rx.try_recv() {
+                let result = match cmd {
+                    WorkerCmd::IbcEvents { batch } => {
+                        // Update scheduled batches.
+                        link.a_to_b.update_schedule(batch)
+                    }
+                    WorkerCmd::NewBlock {
+                        height,
+                        new_block: _,
+                    } => link.a_to_b.clear_packets(height),
+                };
+
+                if let Err(e) = result {
+                    error!("{}", e);
+                    continue;
+                }
+            }
+
+            // Refresh the scheduled batches and execute any outstanding ones.
+            let result = link
+                .a_to_b
+                .refresh_schedule()
+                .and_then(|_| link.a_to_b.execute_schedule());
+
+            if let Err(e) = result {
+                error!("{}", e);
+            }
+        }
+    }
+}
